Adsense teste implementation.

+++ src/pages/JoinTable/ExistentTable.js
@@ -6,12 +6,12 @@ import { FaTrash } from "react-icons/fa";
 import {
     BrowserRouter as Router,
     Link
-  } from "react-router-dom";
+} from "react-router-dom";
 
 export default function ExistentTable() {
 
     const UserId = localStorage.getItem(['user'])
-    const [tables, setTables] = useState(["You've never even create a Table!"])
+    const [tables, setTables] = useState([])
 
     useEffect(() => {
         async function loadTables() {
@@ -28,16 +28,31 @@ export default function ExistentTable() {
         <>
             {
                 tables.map((data, index) => {
-                    return <div className="tableJoin" key={index}>
-                    <Link to="/sessions" ><button className='btn newTableBtn'  onClick={() => {
-                        localStorage.setItem('idTable', data._id)
-                    }} >{data.tableName}</button></Link>
-                    <a onClick={async function(){
-                        let _id = data._id;
-                        const responseDelete = await api.delete('/delete', {headers: {_id}})
-                        console.log(responseDelete);
-                    }}><FaTrash /></a>
-                    </div>
+                    if (data) {
+                        return <div className="tableJoin" key={index}>
+                            <Link to="/sessions" ><button className='btn newTableBtn' onClick={() => {
+                                localStorage.setItem('idTable', data._id)
+                            }} >{data.tableName}</button></Link>
+                            <a onClick={async function () {
+                                let _id = data._id;
+                                const responseDelete = await api.delete('/delete', { headers: { _id } })
+                                console.log(responseDelete);
+                            }}><FaTrash /></a>
+                        </div>
+                    } else {
+                        return <div>
+                            <script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"></script>
+                                <ins class="adsbygoogle"
+                                style="display:block"
+                                data-ad-client="ca-pub-2485166822479834"
+                                data-ad-slot="2411642128"
+                                data-ad-format="auto"
+                                data-full-width-responsive="true"></ins>
+                            <script>
+                                (adsbygoogle = window.adsbygoogle || []).push({});
+                            </script>
+                        </div>
+                    }
                 })
             }
         </>
